Restructures /src files and updates namespaces in class files but not all references. Signed-off-by: Joey Smith <jsmith@webinertia.net>
