+++ src/Pdo/Pdo.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace PhpDb\Adapter\Mysql\Driver\Pdo;
+namespace PhpDb\Mysql\Pdo;
 
 use Override;
 use PDOStatement;
 use PhpDb\Adapter\Driver\Pdo\AbstractPdo;
 use PhpDb\Adapter\Driver\Pdo\Result;
 use PhpDb\Adapter\Driver\ResultInterface;
-use PhpDb\Adapter\Mysql\DatabasePlatformNameTrait;
+use PhpDb\Mysql\DatabasePlatformNameTrait;
 
-class Pdo extends AbstractPdo
+class Driver extends AbstractPdo
 {
     use DatabasePlatformNameTrait;
